немножко рефактора, ща буду смотреть чё с ssl делать
